Refactor: 얼굴 탐지관련 기능 모듈화 진행
